Change counter type

+++ test/data/test_data_times.c
@@ -47,7 +47,7 @@ iter_count(double t, _Complex double v, void *op_data)
 	(void)t;
 	(void)v;
 
-	int *count = op_data;
+	size_t *count = op_data;
 	(*count)++;
 
 	return 0;
@@ -59,7 +59,7 @@ iter_count_onlytwo(double t, _Complex double v, void *op_data)
 	(void)t;
 	(void)v;
 
-	int *count = op_data;
+	size_t *count = op_data;
 	(*count)++;
 	if (*count == 2)
 		return 77;
@@ -102,13 +102,13 @@ test_iter0(void)
 		goto err;
 	}
 
-	int count = 0;
+	size_t count = 0;
 	if (data2_times_foreach(fid, iter_count, &count) != 0) {
 		TEST_FAIL("iteration terminated early");
 		goto err;
 	}
 	if (count != td.num_steps) {
-		TEST_FAIL("number of iterations: %d", count);
+		TEST_FAIL("number of iterations: %zu", count);
 		goto err;
 	}
 
@@ -118,7 +118,7 @@ test_iter0(void)
 		goto err;
 	}
 	if (count != 2) {
-		TEST_FAIL("number of iterations: %d", count);
+		TEST_FAIL("number of iterations: %zu", count);
 		goto err;
 	}
 
@@ -171,13 +171,13 @@ test_iter1(void)
 		goto err;
 	}
 
-	int count = 0;
+	size_t count = 0;
 	if (data2_times_foreach(fid, iter_count, &count) != 0) {
 		TEST_FAIL("iteration terminated early");
 		goto err;
 	}
 	if (count != td.num_steps) {
-		TEST_FAIL("number of iterations: %d", count);
+		TEST_FAIL("number of iterations: %zu", count);
 		goto err;
 	}
 
@@ -187,7 +187,7 @@ test_iter1(void)
 		goto err;
 	}
 	if (count != 2) {
-		TEST_FAIL("number of iterations: %d", count);
+		TEST_FAIL("number of iterations: %zu", count);
 		goto err;
 	}
